requirements.txt
cryptography>=41.0.0
paramiko>=3.3.0
pyjwt>=2.8.0
bcrypt>=4.0.1
argon2-cffi>=23.1.0
click>=8.1.7
rich>=13.6.0
requests>=2.31.0
customtkinter>=6.6.3
packaging>=24.1